requirements.txt
pandas
xlrd
pytorch
torchvision
matplotlib
scikit-image